phpunit tests run! also moving files for capitalization
