+++ tests/test-timber-admin.php
@@ -1,5 +1,7 @@
 <?php
 
+use Timber\Admin;
+
 class TestTimberAdmin extends Timber_UnitTestCase {
 
 	function testSettingsLinks() {
@@ -15,7 +17,7 @@ class TestTimberAdmin extends Timber_UnitTestCase {
     }
 
     function testAdminInit() {
-    	$admin = TimberAdmin::init();
+    	$admin = Admin::init();
     	$this->assertTrue($admin);
     }
